Fix handling of non-square pixels, part 2.
We use CVT registers for inter-glyph communication between the components of
composite glyphs.  However, CVT values might be scaled while reading and
storing to compensate for the aspect ratio between horizontal and vertical
resolution in case of non-square pixels.  Since this scaling is dependent on
the implementation in the bytecode engine, there is no guarantee that a
value stored with WCVTP is the same if read back with RCVT, due to possible
rounding.  In particular, value 1 might be rounded down to value 0.

For this reason, we now use the much larger value 100.