Now switch delta exceptions parser to bison/flex combination.
The most important difference between the old and the new parser is that the
former was line-oriented while the latter is token-oriented (which is
better).  Additionally, it's very easy to extend the grammar now, something
we have to do very soon.

Most of the new code in `TA_deltas_parse_buffer' deals with nice error
message formatting, BTW.