Document Devanagari support.
This is tricky.  Pandoc doesn't support language tags, and inserting special
LaTeX code into the documentation source files is ugly.  For this reason, we
use the `ucharclasses' package so that a proper Devanagari font gets
selected as soon as a character from the Devanagari Unicode block is
present.  Unfortunately, this package currently exists for XeTeX only.