Translate changes to `ta_latin_hint_edges' into bytecode.
It's depressing to see that 3 changed C code lines in this function result
in a few hundred code lines for the bytecode...

We add some higher-level functions, using `top_to_bottom_hinting' as a flag
that gets passed to set the `sal_top_to_bottom_hinting' storage area
location.  The low-level action functions then select the proper comparison
operator depending on this flag.

This is quite repetitive code.