Fix segment handling in bytecode.
Since 2016-07-04 we skip non-edge one-point segments.  However, the
`non-edge' property of such segments is dependent on the PPEM value; it is
thus variable.  This patch introduces a mapping between segments as used in
the auto-hinter structures and the indices used in the bytecode; it is
derived from the lowest PPEM value used for hinting.