Fix bytecode handling of composites with composite subglyphs.
For `TA_font_build_subglyph_shifter', subglyph records must not be further
expanded.

Problem reported by Pablo Moleri <PMoleri@infragistics.com>.  Note that this
bug was there from the very beginning, AFAICS.  FreeType simply ignores such
invalid bytecode (which tried to access a contour with an out-of-bounds
index), but other applications are not that robust.