[doc] Use Noto sans-serif fonts for script documentation.
We are going to document the recently added scripts, for which fonts are
rare to find.  Additionally, it looks more pleasing if all scripts are
represented with the same font family.

This commit also fixes minor glitches (for example, Hebrew wasn't explicitly
handled in `template.tex').

The patch applied to `ucharclasses.sty' has already been added to the
upstream git repository.