Fix compilation on OS X 10.10.
This compiler needs compilation flags to activate C11, which is needed for
gnulib's `max_align_t' implementation.

Note that this is a temporary solution.  The proper, future-proof way
(namely to use two separate `configure' scripts for `lib's C code and
`frontend's C++ code) is outlined in the following two e-mails.

  http://lists.gnu.org/archive/html/bug-gnulib/2017-09/msg00018.html
  http://lists.gnu.org/archive/html/bug-gnulib/2017-09/msg00020.html