[doc] Activate display of Nyiakeng Puachue Hmong.
We need a not-yet-released version of `fontspec.sty`.  Note also that there
doesn't yet exist a Noto sans-serif font for this script, and the serif
version was added to the 'noto-fonts' git repository on 2020-Jun-23, which
is fairly recent.  However, using a serif font is still better than no font.